Chat_V3/ChatApp/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
beautifulsoup4==4.11.1
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
Django==4.1.5
django-bootstrap4==22.3
django-crispy-forms==1.14.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pytz==2022.7
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
zope.interface==5.5.2
